update LayoutProps import path
