+++ .gitignore
@@ -2,4 +2,4 @@ node_modules
 .env
 .next/
 node_modules/
-
+.vscode
